requirements.txt
h5py==2.10.0
numpy==1.18.5
pandas==1.1.5
pathos==0.2.8
potpourri3d==0.0.4
protobuf==3.19.1
pyglet==1.5.21
PyOpenGL==3.1.0
pyrender==0.1.45
Rtree==0.9.7
scipy==1.5.4
Shapely==1.8.0
tensorflow-gpu==2.3.1
trimesh==3.9.35
python-fcl
opencv-python==4.5.4.58
pyYAML==6.0
